ftraceparser: avoid unecessary floating point operations
It doesn't seem to make any difference on x86 but let's avoid
it anyway.

+++ task.h
@@ -30,7 +30,7 @@ public:
 	QVector<double> data;
 	QVector<double> t;
 	QVector<double> scaledData;
-	double lastT;
+	unsigned long long lastT;
 };
 
 #endif
